api/requirements.txt
Django
djangorestframework
drf-spectacular
psycopg2-binary